Extracted common jsonld methods to helper class
