+++ expansion/src/main/java/no/unit/nva/expansion/model/ExpandedParentPublication.java
@@ -2,11 +2,11 @@ package no.unit.nva.expansion.model;
 
 import static no.unit.nva.expansion.ExpansionConfig.objectMapper;
 import static no.unit.nva.expansion.model.ExpandedResource.extractPublicationContextUris;
+import static no.unit.nva.expansion.utils.JsonLdDefaults.frameJsonLd;
 import static nva.commons.apigateway.MediaTypes.APPLICATION_JSON_LD;
 import static nva.commons.core.ioutils.IoUtils.stringToStream;
 import static org.apache.jena.rdf.model.ModelFactory.createDefaultModel;
 import com.fasterxml.jackson.core.JsonProcessingException;
-import com.github.jsonldjava.core.JsonLdOptions;
 import java.io.InputStream;
 import java.net.URI;
 import java.nio.file.Path;
@@ -18,11 +18,8 @@ import java.util.stream.Collectors;
 import no.unit.nva.publication.external.services.UriRetriever;
 import nva.commons.core.ioutils.IoUtils;
 import org.apache.jena.rdf.model.Model;
-import org.apache.jena.riot.JsonLDWriteContext;
 import org.apache.jena.riot.Lang;
 import org.apache.jena.riot.RDFDataMgr;
-import org.apache.jena.riot.RDFFormat;
-import org.apache.jena.riot.RDFWriter;
 import org.apache.jena.vocabulary.RDF;
 
 public class ExpandedParentPublication {
@@ -44,32 +41,7 @@ public class ExpandedParentPublication {
         var model = createDefaultModel();
         loadPublicationWithChannelDataIntoModel(publicationResponseBody.get(), model);
         removePublicationTypeFromResource(publicationId, model);
-        return getFramedModelJson(model);
-    }
-
-    private static String getFramedModelJson(Model model) {
-        return RDFWriter.create()
-                   .format(RDFFormat.JSONLD10_FRAME_PRETTY)
-                   .context(getJsonLdWriteContext())
-                   .source(model)
-                   .build()
-                   .asString();
-    }
-
-    private static JsonLDWriteContext getJsonLdWriteContext() {
-        var context = new JsonLDWriteContext();
-        context.setOptions(getJsonLdOptions());
-        context.setFrame(FRAME);
-        return context;
-    }
-
-    private static JsonLdOptions getJsonLdOptions() {
-        var jsonLdOptions = new JsonLdOptions();
-        jsonLdOptions.setOmitGraph(true);
-        jsonLdOptions.setOmitDefault(true);
-        jsonLdOptions.setUseNativeTypes(true);
-        jsonLdOptions.setPruneBlankNodeIdentifiers(true);
-        return jsonLdOptions;
+        return frameJsonLd(model, FRAME);
     }
 
     private static void removePublicationTypeFromResource(URI id, Model model) {
