+++ expansion/src/main/java/no/unit/nva/expansion/utils/FramedJsonGenerator.java
@@ -1,18 +1,15 @@
 package no.unit.nva.expansion.utils;
 
 import static java.util.Objects.isNull;
-import com.github.jsonldjava.core.JsonLdOptions;
+import static no.unit.nva.expansion.utils.JsonLdDefaults.frameJsonLd;
 import java.io.InputStream;
 import java.util.List;
 import nva.commons.core.JacocoGenerated;
 import org.apache.jena.query.QueryExecutionFactory;
 import org.apache.jena.rdf.model.Model;
 import org.apache.jena.rdf.model.ModelFactory;
-import org.apache.jena.riot.JsonLDWriteContext;
 import org.apache.jena.riot.Lang;
 import org.apache.jena.riot.RDFDataMgr;
-import org.apache.jena.riot.RDFFormat;
-import org.apache.jena.riot.RDFWriter;
 import org.apache.jena.riot.RiotException;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -20,12 +17,11 @@ import org.slf4j.LoggerFactory;
 @JacocoGenerated
 public class FramedJsonGenerator {
     private static final Logger logger = LoggerFactory.getLogger(FramedJsonGenerator.class);
-    private final Model model;
     private final String framedJson;
 
     public FramedJsonGenerator(List<InputStream> streams, String frame) {
-        model = createModel(streams);
-        framedJson = getFramedModelJson(frame);
+        var model = createModel(streams);
+        framedJson = frameJsonLd(model, frame);
     }
 
     public String getFramedJson() {
@@ -57,31 +53,6 @@ public class FramedJsonGenerator {
             var topLevelNode = qexec.execConstruct();
             return model.union(topLevelNode);
         }
-    }
-
-    private String getFramedModelJson(String frame) {
-        return RDFWriter.create()
-                .format(RDFFormat.JSONLD10_FRAME_PRETTY)
-                .context(getJsonLdWriteContext(frame))
-                .source(model)
-                .build()
-                .asString();
-    }
-
-    private static JsonLDWriteContext getJsonLdWriteContext(String frame) {
-        var context = new JsonLDWriteContext();
-        context.setOptions(getJsonLdOptions());
-        context.setFrame(frame);
-        return context;
-    }
-
-    private static JsonLdOptions getJsonLdOptions() {
-        var jsonLdOptions = new JsonLdOptions();
-        jsonLdOptions.setOmitGraph(true);
-        jsonLdOptions.setOmitDefault(true);
-        jsonLdOptions.setUseNativeTypes(true);
-        jsonLdOptions.setPruneBlankNodeIdentifiers(true);
-        return jsonLdOptions;
     }
 
     private static void logInvalidJsonLdInput(Exception exception) {
