+++ expansion/src/main/java/no/unit/nva/expansion/utils/NviCalculator.java
@@ -2,10 +2,10 @@ package no.unit.nva.expansion.utils;
 
 import static java.util.Objects.isNull;
 import static no.unit.nva.expansion.ExpansionConfig.objectMapper;
+import static no.unit.nva.expansion.utils.JsonLdDefaults.frameJsonLd;
 import static nva.commons.core.attempt.Try.attempt;
 import static nva.commons.core.ioutils.IoUtils.stringToStream;
 import com.fasterxml.jackson.core.JsonProcessingException;
-import com.github.jsonldjava.core.JsonLdOptions;
 import java.io.InputStream;
 import java.nio.file.Path;
 import java.util.List;
@@ -15,11 +15,8 @@ import nva.commons.core.ioutils.IoUtils;
 import org.apache.jena.query.QueryExecutionFactory;
 import org.apache.jena.rdf.model.Model;
 import org.apache.jena.rdf.model.ModelFactory;
-import org.apache.jena.riot.JsonLDWriteContext;
 import org.apache.jena.riot.Lang;
 import org.apache.jena.riot.RDFDataMgr;
-import org.apache.jena.riot.RDFFormat;
-import org.apache.jena.riot.RDFWriter;
 import org.apache.jena.riot.RiotException;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -34,14 +31,15 @@ public final class NviCalculator {
         IoUtils.stringFromResources(Path.of("nviTypeQuery.sparql")).replace(NVI_YEAR_REPLACE_STRING, NVI_YEAR);
     private static final Logger logger = LoggerFactory.getLogger(NviCalculator.class);
 
-    private static final String FRAME_SRC = IoUtils.stringFromResources(Path.of("expandedPublicationWithNviFrame.json"));
+    private static final String FRAME_SRC = IoUtils.stringFromResources(
+        Path.of("expandedPublicationWithNviFrame.json"));
 
     private NviCalculator() {
     }
 
     public static ExpandedResource calculateNviType(ExpandedResource expandedResource) throws JsonProcessingException {
         var model = createModelWithNviType(expandedResource);
-        var expandedResourceJsonWithNviType = getFramedModelJson(model);
+        var expandedResourceJsonWithNviType = frameJsonLd(model, FRAME_SRC);
         return attempt(
             () -> objectMapper.readValue(expandedResourceJsonWithNviType, ExpandedResource.class)).orElseThrow();
     }
@@ -64,31 +62,6 @@ public final class NviCalculator {
             var nviType = qexec.execConstruct();
             model.add(nviType);
         }
-    }
-
-    private static String getFramedModelJson(Model model) {
-        return RDFWriter.create()
-                   .format(RDFFormat.JSONLD10_FRAME_PRETTY)
-                   .context(getJsonLdWriteContext())
-                   .source(model)
-                   .build()
-                   .asString();
-    }
-
-    private static JsonLDWriteContext getJsonLdWriteContext() {
-        var context = new JsonLDWriteContext();
-        context.setOptions(getJsonLdOptions());
-        context.setFrame(FRAME_SRC);
-        return context;
-    }
-
-    private static JsonLdOptions getJsonLdOptions() {
-        var jsonLdOptions = new JsonLdOptions();
-        jsonLdOptions.setOmitGraph(true);
-        jsonLdOptions.setOmitDefault(true);
-        jsonLdOptions.setUseNativeTypes(true);
-        jsonLdOptions.setPruneBlankNodeIdentifiers(true);
-        return jsonLdOptions;
     }
 
     @JacocoGenerated
